+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Changelog
 
 ### New
 
+- Support `Map<K, Lazy<V>>` and `Map<K, Provider<Lazy<V>>` multibindings expressions.
+
 ### Enhancements
 
 - [IR] Restructure graph validation and generation to be separate phases, allowing for whole-graph validation before any code gen runs and better optimizing shared bindings across graph extension hierarchies.
@@ -67,6 +69,7 @@ Changelog
 - [IR] Already mentioned above, but worth calling out again — creator-less scoped graph extensions _are no longer cached_ in their parent graphs. Accessors to this will always get new instances now.
 - [IR] Report log files reported from within graph generation now use snake-cased fully-qualified names of the impl graph as the file name suffix.
 - [IR] Do not report similar bindings when a missing binding has type `kotlin.Any`. In practice this reported all available bindings.
+- [interop-dagger] Update to Dagger `2.58`.
 - [Docs] Publish kdocs for guice/jakarta/javax interop and metrox artifacts to doc site.
 - Test Kotlin `2.3.20-Beta1`.